// frontend/src/api/api8082/models/Book.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Authorization server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Book
 */
export interface Book {
    /**
     * 
     * @type {string}
     * @memberof Book
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof Book
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof Book
     */
    author?: string;
    /**
     * 
     * @type {string}
     * @memberof Book
     */
    ownerId?: string;
}

/**
 * Check if a given object implements the Book interface.
 */
export function instanceOfBook(value: object): value is Book {
    return true;
}

export function BookFromJSON(json: any): Book {
    return BookFromJSONTyped(json, false);
}

export function BookFromJSONTyped(json: any, ignoreDiscriminator: boolean): Book {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'title': json['title'] == null ? undefined : json['title'],
        'author': json['author'] == null ? undefined : json['author'],
        'ownerId': json['ownerId'] == null ? undefined : json['ownerId'],
    };
}

export function BookToJSON(json: any): Book {
    return BookToJSONTyped(json, false);
}

export function BookToJSONTyped(value?: Book | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'title': value['title'],
        'author': value['author'],
        'ownerId': value['ownerId'],
    };
}