// frontend/src/api/api8082/models/BookDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Authorization server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BookDto
 */
export interface BookDto {
    /**
     * 
     * @type {string}
     * @memberof BookDto
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof BookDto
     */
    author?: string;
}

/**
 * Check if a given object implements the BookDto interface.
 */
export function instanceOfBookDto(value: object): value is BookDto {
    return true;
}

export function BookDtoFromJSON(json: any): BookDto {
    return BookDtoFromJSONTyped(json, false);
}

export function BookDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): BookDto {
    if (json == null) {
        return json;
    }
    return {
        
        'title': json['title'] == null ? undefined : json['title'],
        'author': json['author'] == null ? undefined : json['author'],
    };
}

export function BookDtoToJSON(json: any): BookDto {
    return BookDtoToJSONTyped(json, false);
}

export function BookDtoToJSONTyped(value?: BookDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'title': value['title'],
        'author': value['author'],
    };
}